pyproject.toml
[tool.poetry]
name = "ai-vm"
version = "0.1.0"
description = ""
authors = ["Matéo Mévollon <m.mevollon@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ovh = "^1.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
